Reorganized and updated code style
