+++ src/main/kotlin/org/radarbase/gateway/io/AvroProcessor.kt
@@ -85,7 +85,8 @@ class AvroProcessor(
 
     private fun createMapping(topic: String, ofValue: Boolean, sourceSchema: Schema): JsonToObjectMapping {
         val latestSchema = schemaRetriever.getBySubjectAndVersion(topic, ofValue, -1)
-        val schemaMapper = AvroDataMapperFactory.get().createMapper(sourceSchema, latestSchema.schema, null)
+        val schemaMapper = AvroDataMapperFactory.get()
+            .createMapper(sourceSchema, latestSchema.schema, null)
         return JsonToObjectMapping(sourceSchema, latestSchema.schema, latestSchema.id, schemaMapper)
     }
